python_core/requirements.txt
ollama
ncps
eclipse-zenoh
pymdp
numpy
torch
websockets
psutil
